Make snippet placeholders syntactically less problematic
The current placeholders wrapped in <> brackets would cause the linter to
believe that all kinds of things were wrong, because < and > are also
binary operators. Furthermore, it actually leads to a JS error when wanting
to autocomplete in the second field of the histogram_quantile() snippet,
but I couldn't track down why yet. It's definitely linter-related though,
as the JS error only happens with the linter enabled. This change at least
avoids that common situation for now.

Signed-off-by: Julius Volz <julius.volz@gmail.com>